Настройка эффектов перенесена в отдельный модуль
